add SpatiumFunctions::with_correction constructor
