+++ src/strength/internal_forces/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod bending_moment;
+pub(crate) mod internal_forces;
 pub(crate) mod share_force;
