+++ src/strength/internal_forces/share_force.rs
@@ -27,7 +27,7 @@ impl<'a> ShareForce<'a> {
         ship_dimensions: &ShipDimensions,
     ) -> Result<SpatiumFunctions, String> {
         let total_shipload = self.total_shipload.total_shipload(ship_dimensions)?;
-        let mut share_force = total_shipload.integral_vul();
+        let share_force = total_shipload.integral_vul();
         let nose_share_force = share_force.last().unwrap().f_x2();
         let max_share_force = share_force.max().unwrap();
         if nose_share_force.abs() / max_share_force > 0.05 {
@@ -36,23 +36,9 @@ impl<'a> ShareForce<'a> {
                 nose_share_force / max_share_force
             );
         }
-        let mut f_x1 = 0.0;
-        let mut s_fs = vec![];
-        let mut x = ship_dimensions.length_spatium();
-        let lbp = ship_dimensions.lbp();
-        for s_f in share_force.into_iter() {
-            let f_x2 = s_f.f_x2() - nose_share_force * x / lbp;
-            s_fs.push(SpatiumFunction::new(
-                s_f.id(),
-                s_f.x1(),
-                s_f.x2(),
-                f_x1,
-                f_x2,
-            ));
-            f_x1 = f_x2;
-            x += ship_dimensions.length_spatium();
-        }
-        share_force = SpatiumFunctions::new(s_fs);
-        Ok(share_force)
+        return Ok(SpatiumFunctions::with_correction(
+            share_force,
+            ship_dimensions,
+        ));
     }
 }
