+++ src/strength/ship/spatium_functions.rs
@@ -1,8 +1,9 @@
+use log::warn;
 use serde::Deserialize;
 
 use crate::core::round::Round;
 
-use super::spatium_function::SpatiumFunction;
+use super::{ship_dimensions::ShipDimensions, spatium_function::SpatiumFunction};
 
 ///
 /// Содержит результаты вычислений
@@ -44,6 +45,51 @@ impl SpatiumFunctions {
     }
 
     ///
+    /// Вспомогательный коструктор.
+    /// Так как удифферентовка судна осуществляется приближенно,
+    /// то после интегрирования суммарной нагрузки в носовом шпангоуте появляются
+    /// невязки в эпюрах: N(nose) != 0, M(nose) = 0. Для получения эпюр N(x), M(x),
+    /// удовлетворяющих граничным условиям, данные невязки линейно разносят по шпациям.
+    /// Исправленные N и M равны:
+    /// N_исп = N(x) - N_nose * x / L, M_исп = M_nose * x / L, где
+    /// x - отсчитывается от кормового перепендикуляра,
+    /// N(x) - перерезывающая сила в точке с координатой x,
+    /// M(x) - изгибающий момент с координатой x,
+    /// N_nose - перерезывающая сила в носовом шпангоуте,
+    /// M_nose - изгибающий момент в носовм шпангоуте,
+    /// L - длина между перпендикулярами.
+    pub fn with_correction(
+        spatium_functions: SpatiumFunctions,
+        ship_dimensions: &ShipDimensions,
+    ) -> SpatiumFunctions {
+        let nose_value = spatium_functions.last().unwrap().f_x2();
+        let max_value = spatium_functions.max().unwrap();
+        if nose_value.abs() / max_value > 0.05 {
+            warn!(
+                "Эпюра не замкнута. Незамыкание эпюры: Value(nose) / Value(max) = {}",
+                nose_value / max_value
+            );
+        }
+        let mut f_x1 = 0.0;
+        let mut s_fs = vec![];
+        let mut x = ship_dimensions.length_spatium();
+        let lbp = ship_dimensions.lbp();
+        for s_f in spatium_functions.into_iter() {
+            let f_x2 = s_f.f_x2() - nose_value * x / lbp;
+            s_fs.push(SpatiumFunction::new(
+                s_f.id(),
+                s_f.x1(),
+                s_f.x2(),
+                f_x1,
+                f_x2,
+            ));
+            f_x1 = f_x2;
+            x += ship_dimensions.length_spatium();
+        }
+        SpatiumFunctions::new(s_fs)
+    }
+
+    ///
     /// Возвращает максимальное значение по модулю.
     pub fn max(&self) -> Option<f64> {
         if self.spatium_functions.len() == 0 {
@@ -60,7 +106,7 @@ impl SpatiumFunctions {
     }
 
     ///
-    /// Вычисляет интеграл с переменным верхним пределом;
+    /// Вычисляет интеграл с переменным верхним пределом.
     pub fn integral_vul(&self) -> SpatiumFunctions {
         let mut spatium_functions = vec![];
         let mut f_x1 = 0.0;
